orientation-event/DeviceMotion/t025.html

<!DOCTYPE html>
<html>

  <head>
	<title>Implementations that are unable to provide acceleration data without the effect of gravity may instead supply the acceleration including the effect of gravity:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
  <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  		<h1>Implementations that are unable to provide acceleration data without the effect of gravity may instead supply the acceleration including the effect of gravity:</h1>
    <p id="eaccgvt"></p>
	<div id="log"></div>
	
	<script type="text/javascript">
		t1 = async_test("Implementation unable to provide 'acceleration' property");
		t2 = async_test("Implementation able to provide 'acceleration IncludingGravity' property");
		eacc = 0;
		eaccgvt = 0;
		window.ondevicemotion = function(e){
			eacc = e.acceleration;
			eaccgvt = e.accelerationIncludingGravity;
		};
		
		setTimeout("testing()", 200);
		
		
		function testing(){
			t1.step(function() {assert_true(eacc == null)});
			t2.step(function() {assert_true((typeof eaccgvt == "object") && ((eaccgvt.x != 0) && (eaccgvt.y != 0) && (eaccgvt.z != 0)))});
			t1.done();
			t2.done();
		}
	</script>	
  </body>
  
</html>
